requirements.txt
asgiref
Django
pytz
sqlparse
djangorestframework
six
django-cors-headers 